+++ cmd/frontend/internal/graphqlbackend/search_repositories.go
@@ -24,7 +24,7 @@ func searchRepositories(ctx context.Context, args *repoSearchArgs, query searchq
 		searchquery.FieldType:      struct{}{},
 		searchquery.FieldDefault:   struct{}{},
 		searchquery.FieldIndex:     struct{}{},
-		searchquery.FieldMax:       struct{}{},
+		searchquery.FieldCount:     struct{}{},
 		searchquery.FieldTimeout:   struct{}{},
 	}
 	// Don't return repo results if the search contains fields that aren't on the whitelist.